Source: https://github.com/msupply-foundation/open-msupply/commit/43e2a723b3f4936eb9778b2b0f8cf2a9d79b1c99

+++ server/graphql/core/src/loader/item_stats.rs
@@ -12,7 +12,7 @@ pub struct ItemsStatsForItemLoader {
     pub service_provider: Data<ServiceProvider>,
 }
 
-pub type ItemStatsLoaderInputPayload = Option<u32>; // amc_lookback_months
+pub type ItemStatsLoaderInputPayload = Option<f64>; // amc_lookback_months
 pub type ItemStatsLoaderInput = IdPair<ItemStatsLoaderInputPayload>;
 impl ItemStatsLoaderInput {
     pub fn new(store_id: &str, item_id: &str, payload: ItemStatsLoaderInputPayload) -> Self {
@@ -24,7 +24,6 @@ impl ItemStatsLoaderInput {
     }
 }
 
-#[async_trait::async_trait]
 impl Loader<ItemStatsLoaderInput> for ItemsStatsForItemLoader {
     type Value = ItemStats;
     type Error = async_graphql::Error;
